Add --fuse option
This allows mounting FUSE filesystems in the sandbox by passing an open
/dev/fuse fd to bwrap.

The kernel currently requires /dev/fuse to be opened in the same user
namespace as it is mounted. With this restriction it's not possible to
combine --fuse with --unshare-user. However, a parent process of bwrap
can unshare the user namespace, open /dev/fuse, spawn bwrap passing the
FUSE file descriptor to --fuse, and then spawn a FUSE process passing
the FUSE file descriptor as special mount point /dev/fd/N.